native: add sample rate conversion for SFX playback
Bank samples are 32kHz but AudioDevice outputs at 44.1kHz. Without
resampling, sounds play 1.38x too fast. Added linear interpolation
with fractional position tracking and rate ratio adjustment.

+++ native/src/platform/SampleInst_Native.h
@@ -33,7 +33,7 @@ private:
     FxSend *mFxSend = nullptr;
     const int16_t *mPCMData;    // pointer into SampleData's buffer
     int mPCMSamples;            // total samples available
-    int mPlayPos;               // current sample position
+    double mPlayPos;            // fractional sample position for resampling
     int mEndSample;             // where to stop (-1 = end of data)
     bool mLoop;
     volatile bool mPlaying;
